refactored DeckboxCloneForm to use TS
